requirements.txt
aniso8601==7.0.0
asgiref==3.4.1
backports.zoneinfo==0.2.1
django<=3
django-graphql-auth== 0.3.16
graphene==2.1.9
django-graphql-jwt==0.3.4
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
promise==2.3
psycopg2-binary==2.8.6
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3